Separated one component into several components
