+++ src/app/(admin)/admin/partners/components/Partners/Partners.module.scss
@@ -1,0 +1,24 @@
+@import '../../../../../../scss/utils/variables';
+
+.wrapper {
+  display: flex;
+  flex-direction: column;
+  height: 100vh;
+  justify-content: space-between;
+
+  .spinner {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    width: 70px;
+    height: 70px;
+  }
+
+  .error {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translateX(-50%);
+    color: $error;
+  }
+}
