+++ src/app/(admin)/admin/partners/components/PartnersList/PartnersList.module.scss
@@ -1,32 +1,7 @@
-@import '../../../../../../scss/utils/variables';
-
 .partners {
   display: flex;
   flex-direction: column;
   gap: 24px;
   position: relative;
   height: 100%;
-
-  .wrapper {
-    display: flex;
-    flex-direction: column;
-    height: 100vh;
-    justify-content: space-between;
-
-    .spinner {
-      position: absolute;
-      top: 50%;
-      left: 50%;
-      width: 70px;
-      height: 70px;
-    }
-
-    .error {
-      position: absolute;
-      top: 50%;
-      left: 50%;
-      transform: translateX(-50%);
-      color: $error;
-    }
-  }
 }
